+++ src/fs/copy.js
@@ -1,9 +1,9 @@
 import * as fs from "node:fs/promises";
-import { createPath, FS_ERROR_MESSAGE } from "../utils.js";
+import { createPath, FS_ERROR_MESSAGE } from "./utils.js";
 
 const copy = async () => {
-  const src = createPath("fs/files");
-  const dest = createPath("fs/files_copy");
+  const src = createPath("files");
+  const dest = createPath("files_copy");
 
   try {
     await fs.cp(src, dest, {recursive:true, force: false, errorOnExist: true});
