+++ src/fs/rename.js
@@ -1,9 +1,9 @@
 import * as fs from "node:fs/promises";
-import { createPath, FS_ERROR_MESSAGE } from "../utils.js";
+import { createPath, FS_ERROR_MESSAGE } from "./utils.js";
 
 const rename = async () => {
-  const oldPath = createPath("fs/files/wrongFilename.txt");
-  const newPath = createPath("fs/files/properFilename.md");
+  const oldPath = createPath("files/wrongFilename.txt");
+  const newPath = createPath("files/properFilename.md");
 
   try {
     await fs.rename(oldPath, newPath);
