+++ src/zip/compress.js
@@ -1,15 +1,15 @@
 import { promisify } from "node:util";
 import { createGzip } from "node:zlib";
-import { createPath } from "../utils.js";
 import { pipeline } from "node:stream";
 import { createReadStream, createWriteStream } from "node:fs";
+import path from 'node:path';
 
 const pipe = promisify(pipeline);
 
 const compress = async () => {
   try {
-    const srcFile = createPath("zip/files/fileToCompress.txt");
-    const destFile = createPath("zip/files/archive.gz");
+    const srcFile = path.join(import.meta.dirname, "files/fileToCompress.txt");
+    const destFile = path.join(import.meta.dirname, "files/archive.gz");
     const gzip = createGzip();
     const source = createReadStream(srcFile);
     const destination = createWriteStream(destFile);
